+++ src/Controller/GaUsuarioController.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Model\Entity\GaNivelUsuario;
+use Cake\Http\Exception\ForbiddenException;
 /**
  * GaUsuario Controller
  *
@@ -13,12 +15,26 @@ namespace App\Controller;
 class GaUsuarioController extends AppController
 {
 
+    public function autorizado(){
+        foreach($this->nivelUsuario as $nivel){
+            if(($nivel->ga_nivel_acesso->sigla == "ADM") OR ($nivel->ga_nivel_acesso->sigla == "GA")){
+                return true;
+                break;
+            } else {
+                return false;
+                break;
+            }
+        }
+    }
+
+
     public function login()
-    {
+    {        //não precisa de autorização para logar
+       //$this->Authorization->skipAuthorization();
         $result = $this->Authentication->getResult();
         // If the user is logged in send them away.
         if ($result->isValid()) {
-            $target = $this->Authentication->getLoginRedirect() ?? '/home';
+            $target = $this->Authentication->getLoginRedirect() ?? '/';
             return $this->redirect($target);
         }
         if ($this->request->is('post') && !$result->isValid()) {
@@ -28,6 +44,8 @@ class GaUsuarioController extends AppController
 
     public function logout()
     {
+        //não precisa de autorização para deslogar
+        //$this->Authorization->skipAuthorization();
         $this->Authentication->logout();
         return $this->redirect(['controller' => 'GaUsuario', 'action' => 'login']);
     }
@@ -40,8 +58,13 @@ class GaUsuarioController extends AppController
      */
     public function index()
     {
-        $gaUsuario = $this->paginate($this->GaUsuario);
-        $this->set(compact('gaUsuario'));
+        if($this->autorizado()){
+            $gaUsuario = $this->paginate($this->GaUsuario);
+            $this->set(compact('gaUsuario'));
+        }else {
+            throw new ForbiddenException(__('Você não possui acesso a este módulo'));
+            //return $this->redirect(['controller' => 'GaUsuario', 'action' => 'logout']);
+        }
     }
 
     /**
@@ -53,11 +76,16 @@ class GaUsuarioController extends AppController
      */
     public function view($id = null)
     {
-        $gaUsuario = $this->GaUsuario->get($id, [
-            'contain' => [],
-        ]);
+        if($this->autorizado()){
+            $gaUsuario = $this->GaUsuario->get($id, [
+                'contain' => [],
+            ]);
 
-        $this->set('gaUsuario', $gaUsuario);
+            $this->set('gaUsuario', $gaUsuario);
+        }else {
+            throw new ForbiddenException(__('Você não possui acesso a este módulo'));
+            //return $this->redirect(['controller' => 'GaUsuario', 'action' => 'logout']);
+        }
     }
 
     /**
@@ -67,17 +95,24 @@ class GaUsuarioController extends AppController
      */
     public function add()
     {
-        $gaUsuario = $this->GaUsuario->newEmptyEntity();
-        if ($this->request->is('post')) {
-            $gaUsuario = $this->GaUsuario->patchEntity($gaUsuario, $this->request->getData());
-            if ($this->GaUsuario->save($gaUsuario)) {
-                $this->Flash->success(__('Usuário salvo com sucesso'));
+        if($this->autorizado()){
+            $gaUsuario = $this->GaUsuario->newEmptyEntity();
+            if ($this->request->is('post')) {
+                $gaUsuario = $this->GaUsuario->patchEntity($gaUsuario, $this->request->getData());
+                if ($this->GaUsuario->save($gaUsuario)) {
+                    $this->Flash->success(__('Usuário salvo com sucesso'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('O usuário não pode ser salvo. Por favor tente novamente!'));
             }
-            $this->Flash->error(__('O usuário não pode ser salvo. Por favor tente novamente!'));
+            $this->set(compact('gaUsuario'));
+        }else {
+            throw new ForbiddenException(__('Você não possui acesso a este módulo'));
+            //return $this->redirect(['controller' => 'GaUsuario', 'action' => 'logout']);
         }
-        $this->set(compact('gaUsuario'));
+
+
     }
 
     /**
@@ -89,19 +124,24 @@ class GaUsuarioController extends AppController
      */
     public function edit($id = null)
     {
-        $gaUsuario = $this->GaUsuario->get($id, [
-            'contain' => [],
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $gaUsuario = $this->GaUsuario->patchEntity($gaUsuario, $this->request->getData());
-            if ($this->GaUsuario->save($gaUsuario)) {
-                $this->Flash->success(__('Usuário editado com sucesso.'));
+        if($this->autorizado()){
+            $gaUsuario = $this->GaUsuario->get($id, [
+                'contain' => [],
+            ]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $gaUsuario = $this->GaUsuario->patchEntity($gaUsuario, $this->request->getData());
+                if ($this->GaUsuario->save($gaUsuario)) {
+                    $this->Flash->success(__('Usuário editado com sucesso.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('O usuário não pode ser salvo. Por favor tente novamente!'));
             }
-            $this->Flash->error(__('O usuário não pode ser salvo. Por favor tente novamente!'));
+            $this->set(compact('gaUsuario'));
+        }else {
+            throw new ForbiddenException(__('Você não possui acesso a este módulo'));
+            //return $this->redirect(['controller' => 'GaUsuario', 'action' => 'logout']);
         }
-        $this->set(compact('gaUsuario'));
     }
 
     /**
@@ -113,14 +153,21 @@ class GaUsuarioController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $gaUsuario = $this->GaUsuario->get($id);
-        if ($this->GaUsuario->delete($gaUsuario)) {
-            $this->Flash->success(__('O usuário foi excluído!'));
-        } else {
-            $this->Flash->error(__('O usuário não pode ser excluído. Por favor tente novamente'));
+        if($this->autorizado()){
+            $this->request->allowMethod(['post', 'delete']);
+            $gaUsuario = $this->GaUsuario->get($id);
+            if ($this->GaUsuario->delete($gaUsuario)) {
+                $this->Flash->success(__('O usuário foi excluído!'));
+            } else {
+                $this->Flash->error(__('O usuário não pode ser excluído. Por favor tente novamente'));
+            }
+
+            return $this->redirect(['action' => 'index']);
+        }else {
+            throw new ForbiddenException(__('Você não possui acesso a este módulo'));
+            //return $this->redirect(['controller' => 'GaUsuario', 'action' => 'logout']);
         }
 
-        return $this->redirect(['action' => 'index']);
+
     }
 }
